refactor to cleaner folders
